refactor: 멀티파트 변환 객체 이동
