+++ app/src/main/java/com/into/websoso/data/mapper/MultiPartMapper.kt
@@ -1,4 +1,4 @@
-package com.into.websoso.core.common.util
+package com.into.websoso.data.mapper
 
 import android.content.Context
 import android.net.Uri
@@ -12,7 +12,7 @@ import okhttp3.RequestBody.Companion.toRequestBody
 import java.io.InputStream
 import javax.inject.Inject
 
-class MultiPartConvertor
+class MultiPartMapper
     @Inject
     constructor(
         @ApplicationContext private val context: Context,
